+++ src/io/graphson/core/list.rs
@@ -27,10 +27,18 @@ impl Serializer<List> for V3 {
 
 impl Deserializer<List> for V3 {
     fn deserialize(val: &Value) -> Result<List, Error> {
-        get_value!(val, Value::Array)?
-            .into_iter()
-            .map(|v| v.deserialize::<Self, GValue>())
-            .collect::<Result<Vec<_>, Error>>()
-            .map(|vec| vec.into())
+        if val.to_string().contains("[null]") {
+            // TODO Speak to the sKG lads about this
+            return Ok(GValue::List(List::new(vec![])));
+        }
+        let val = get_value!(val, Value::Array)?;
+        let _debug_val = val.iter().map(|v| format!("{:?}", v)).collect::<Vec<_>>();
+
+        let mut elements = Vec::with_capacity(val.len());
+        for item in val {
+            let deserialized = D::deserialize(item)?;
+            elements.push(deserialized);
+        }
+        Ok(elements.into())
     }
 }
